+++ crates/zune-imageprocs/src/unsharpen.rs
@@ -1,8 +1,8 @@
 use crate::gaussian_blur::gaussian_blur;
 
 pub fn unsharpen(
-    channel: &mut [u8], blur_buffer: &mut [u8], blur_scratch_buffer: &mut [u8], sigma: f32,
-    threshold: u8, width: usize, height: usize,
+    channel: &mut [u16], blur_buffer: &mut [u16], blur_scratch_buffer: &mut [u16], sigma: f32,
+    threshold: u16, width: usize, height: usize,
 )
 {
     // copy channel to scratch space
@@ -20,7 +20,7 @@ pub fn unsharpen(
         // We conditionally take the added version or whatever we had based on this mask
         //  godbolt link: https://godbolt.org/z/YYnEaPedM
 
-        let threshold_mask = u8::from(diff > threshold).wrapping_sub(1);
+        let threshold_mask = u16::from(diff > threshold).wrapping_sub(1);
 
         // if diff > threshold { pix = (diff + pix) } else { pix }
         *in_pix = (in_pix.saturating_add(diff) & !threshold_mask) | (*in_pix & threshold_mask);
